Implement File::replace_contents_async manually
The contents need to be owned until the callback has triggered.
